fix: don't call transformResponses for components
transformResponses ignores properties which are not valid HTTP response
codes (other than default), since they are not defined on the Responses
Object".  This is a problem, since components.responses is not a
"Responses Object", but a "Map[string,Response Object]" where property
names can be anything.  Therefore, call transformMap instead.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>