add-tag-to-operation-ids constructor options
Rework constructor options handling to be more concise.
Unlike the previous code, it now throws TypeError for null.
Unlike other classes, it does not throw for non-object values.

I'm on the fence about whether to type check the options argument.
Informing callers they have made an error is useful, but it's more code
to maintain and is at odds with most of the JS ecosystem.  Probably
better to add TypeScript for users who want type errors.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>