rename-components: skip recursive transform $ref
renameRefObj served the dual purpose of identifying JSON References and
applying the rename if found.  However, returning falsey for non-refs
is not intuitive and wasn't applied correctly if no renames were
applicable, causing recursion to be attempted on JSON Reference objects.

Split the logic for identifying JSON Reference objects from the rename
logic to make this clearer and avoid unnecessary traversal.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>