fix(rename-components): tighten $ref matching
The RegExp to limit $ref paths was not being used, and would not have
handled percent-encoding correctly anyway.  The renaming always applied
to the last path component, which was incorrect for references to child
schemas.

Fix these issues by only renaming the path segment returned by a
function which matches the JSON path.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>